Switch from anon inodes to our own
Since we own the inode, we can configure it as we like.  Most
importantly, we can mark it as a character device.

This is to ensure that standard strems buffering is configured exactly
the same as when running interactively.  At least glibc is known to
check inode type.  Python does a moral equivalent of isatty(), which was
already handled.